requirements.txt
django
django-cors-headers
python-dotenv
openai
anthropic
mistralai
sacrebleu